Fix writing of encoder settings to sensor
This patch fixes an issue while writing the enconder settings to sensors.
When the value is left default, the value written to the sensor is 1.
Writing 1 to lms15xx will enable output of encoder values in datagram,
and thus making the datagram parser to fail due to the additional data.